pyproject.toml
[project]
name = "nonebot-webui-admin"
version = "0.1.0"
description = "WebUI Management System for NoneBot"
dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-adapter-onebot>=2.0.0",
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "sqlalchemy>=1.4.0",
    "aiosqlite>=0.17.0",
    "pydantic>=1.8.0",
    "jinja2>=3.0.0",
    "python-multipart>=0.0.5",
    "aiofiles>=0.7.0"
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
builtin_plugins = ["echo"]

[project.scripts]
webui-admin = "main:main"